Improved density threshold. Added point_cloud actor.
